+++ Front/app/src/main/java/com/example/capston_spotyup/Signup/Domain/fragmentInfo.kt
@@ -1,15 +1,12 @@
 package com.example.capston_spotyup.Signup.Domain
 
 import android.content.Intent
-import android.graphics.Color
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentTransaction
-import com.example.capston_spotyup.Main.MainActivity
 import com.example.capston_spotyup.R
 import com.example.capston_spotyup.Signin.Domain.SigninActivity
 import com.example.capston_spotyup.databinding.SignupInfoBinding
